fix(ci): frontend structure — primitive controls, ProviderRow extraction, editor ratchet entry
RAW_DOM_CONTROL: ProviderRow's raw <button> and ProviderConfigPanel's raw
<label> now use the Button/Label primitives. LARGE_FRONTEND_FILE:
ProviderRow (+ProviderLogo) extracted from ProviderPickerModal (411→272
lines); use-harness-auth-editor.ts recorded in the ratchet allowlist at
410 pending a split.
