+++ scripts/frontend_structure_allowlist.txt
@@ -38,3 +38,4 @@ apps/packages/product-client/src/components/playground/WorkspaceStatusPlayground
 apps/packages/product-client/src/components/playground/GitReviewV2Playground.tsx 445 git review playground fixtures pending scenario-module split
 apps/packages/product-client/src/hooks/plans/workflows/use-proposed-plan-actions.ts 405 plan decision + carry-out workflow growth (#1250); split deferred
 apps/packages/design/src/tokens.ts 2143 single token authority: one entry per token with provenance by design; shape is a flat manifest, not logic
+apps/packages/product-client/src/hooks/agents/workflows/use-harness-auth-editor.ts 410 harness auth editor grew with v2 single-PUT replace semantics + provider-hint rows; split pending
